Separate logic to utilities file
